Rename `abstract_level.py` back to `level.py`
