+++ src/finch/tensor/level/__init__.py
@@ -1,6 +1,6 @@
-from .abstract_level import Level, LevelFType
 from .dense_level import DenseLevel, DenseLevelFType, dense
 from .element_level import ElementLevel, ElementLevelFType, element
+from .level import Level, LevelFType
 from .sparse_bytemap_level import (
     SparseByteMapLevel,
     SparseByteMapLevelFType,
