+++ src/finch/tensor/level/sparse_list_level.py
@@ -9,9 +9,8 @@ from finch.algebra import FType, ImmutableStructFType, ffuncs, ftype, ftypes
 from finch.compile import looplets as lplt
 from finch.finch_assembly import parse_assembly
 from finch.tensor.fiber_tensor import FiberTensorFType
+from finch.tensor.level import Level, LevelFType
 from finch.tensor.scalar import Scalar
-
-from .abstract_level import Level, LevelFType
 
 
 @dataclass(unsafe_hash=True)
